Separate main.go (#12)
* follow ddd

* add searcher text

* add domain/path tests

* refactor presentation/path.go

* add presentation/path tests

* remove main tests
