+++ domain/searcher/searcher_suite_test.go
@@ -1,0 +1,13 @@
+package searcher_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+func TestSearcher(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Searcher Suite")
+}
