+++ presentation/path/path_suite_test.go
@@ -1,0 +1,13 @@
+package path_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+func TestPath(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Path Suite")
+}
